+++ qCali/src/main/java/com/group/exam/board/dao/BoardDao.java
@@ -18,21 +18,21 @@ public interface BoardDao {
 
 	public void updateBoard(HashMap<String, Object> map); // 게시글 수정
 
-	public void deleteBoardOne(HashMap<String, Integer> map); // 게시글 삭제
+	public void deleteBoardOne(HashMap<String, Long> map); // 게시글 삭제
 
 	public List<BoardlistCommand> boardList(Criteria cri); // 게시글 전체 리스트
 
 	public List<BoardlistCommand> boardMyList(HashMap<String, Object> map); // 내가 쓴 글 모아보기
 
-	public BoardlistCommand boardListDetail(int boardSeq); // 특정 게시글 디테일
+	public BoardlistCommand boardListDetail(Long boardSeq); // 특정 게시글 디테일
 
 	public int listCount(); // board 테이블 전체 글 수
 
-	public int boardMylistCount(int memberSeq); // 내가 쓴 글 수
+	public int boardMylistCount(Long memberSeq); // 내가 쓴 글 수
 
-	public void boardCountup(int boardSeq); // 해당 게시글 카운트 업
+	public void boardCountup(Long boardSeq); // 해당 게시글 카운트 업
 
-	public String memberAuth(int memberSeq); // 멤버 Auth 상태 체크
+	public String memberAuth(Long memberSeq); // 멤버 Auth 상태 체크
 
 	public int memberLevelup(HashMap<String, Object> map); // 멤버 level up 기능
 
@@ -52,19 +52,19 @@ public interface BoardDao {
 
 	public void deleteBoardLike(BoardHeartVo vo);
 
-	public void updateBoardLike(int boardSeq);
+	public void updateBoardLike(Long boardSeq);
 
 	// 댓글 기능 관련
-	public List<ReplyVo> replySelect(int boardSeq); // 게시글에 맞춰서 댓글 리스트 띄우기
+	public List<ReplyVo> replySelect(Long boardSeq); // 게시글에 맞춰서 댓글 리스트 띄우기
 
-	public int replyCount(int boardSeq); // 댓글 갯수 카운트 -> 해당 글에 달린 댓글 수를 가져와야함.
+	public int replyCount(Long boardSeq); // 댓글 갯수 카운트 -> 해당 글에 달린 댓글 수를 가져와야함.
 
 	public void replyInsert(ReplyVo replyVo); // 댓글 쓰기
 
 	public void replyUpdate(ReplyVo replyVo); // 댓글 수정
 
-	public void replyDelete(int replySeq); // 댓글 삭제
-	
+	public void replyDelete(Long replySeq); // 댓글 삭제
+
 	// 현재 시퀀스 구하기(insert 뒤에만 적용할 예정이라 current 사용 가능)
 	public int currentBoardSeq();
