+++ qCali/src/main/java/com/group/exam/board/service/BoardService.java
@@ -17,23 +17,23 @@ public interface BoardService {
 
 	public List<BoardlistCommand> boardList(Criteria cri); // 게시글 전체 리스트
 
-	public List<BoardlistCommand> boardMyList(Criteria cri, int memberSeq); // 내가 쓴 글 불러오기
+	public List<BoardlistCommand> boardMyList(Criteria cri, Long memberSeq); // 내가 쓴 글 불러오기
 
-	public BoardlistCommand boardListDetail(int boardSeq); // 해당 게시글 디테일
+	public BoardlistCommand boardListDetail(Long boardSeq); // 해당 게시글 디테일
 
-	public void boardCountup(int boardSeq); // 해당 게시글 카운트 업
+	public void boardCountup(Long boardSeq); // 해당 게시글 카운트 업
 
-	public void deleteBoardOne(int boardSeq, int memberSeq); // 해당 게시글 삭제
+	public void deleteBoardOne(Long boardSeq, Long memberSeq); // 해당 게시글 삭제
 
 	public int listCount(); // board 테이블 전체 글 수
 
-	public int mylistCount(int memberSeq); // 게시글 쓴 글 수
+	public int mylistCount(Long memberSeq); // 게시글 쓴 글 수
 
-	public void updateBoard(String boardTitle, String boardContent, int boardSeq); // 게시글 수정
+	public void updateBoard(String boardTitle, String boardContent, Long boardSeq); // 게시글 수정
 
-	public String memberAuth(int memberSeq); // 멤버 Auth 상태 확인
+	public String memberAuth(Long memberSeq); // 멤버 Auth 상태 확인
 
-	public int memberLevelup(int memberSeq, int mytotal, int memberLevel);
+	public int memberLevelup(Long memberSeq, int mytotal, int memberLevel);
 
 	public List<BoardlistCommand> boardSearch(HashMap<String, Object> map);// 닉네임, 제목으로 검색
 
@@ -53,13 +53,13 @@ public interface BoardService {
 	public int getBoardLike(BoardHeartVo vo);
 
 	// 댓글 기능 관련
-	public List<ReplyVo> replySelect(int boardSeq); // 게시글에 맞춰서 댓글 리스트 띄우기
+	public List<ReplyVo> replySelect(Long boardSeq); // 게시글에 맞춰서 댓글 리스트 띄우기
 
 	public void replyInsert(ReplyVo replyVo); // 댓글 쓰기
 
 	public void replyUpdate(ReplyVo replyVo); // 댓글 수정
 
-	public void replyDelete(int replySeq); // 댓글 삭제
+	public void replyDelete(Long replySeq); // 댓글 삭제
 	
 	// 현재 시퀀스 구하기(insert 뒤에만 적용할 예정이라 current 사용 가능)
 	public int currentBoardSeq();
