+++ qCali/src/main/java/com/group/exam/notice/command/NoticeUpdateCommand.java
@@ -7,33 +7,36 @@ public class NoticeUpdateCommand {
 	private String noticeTitle;
 	@NotEmpty
 	private String noticeContent;
-	private int noticeSeq;
-	
-	
-	public int getNoticeSeq() {
+	private Long noticeSeq;
+
+	public Long getNoticeSeq() {
 		return noticeSeq;
 	}
-	public void setNoticeSeq(int noticeSeq) {
+
+	public void setNoticeSeq(Long noticeSeq) {
 		this.noticeSeq = noticeSeq;
 	}
+
 	public String getNoticeTitle() {
 		return noticeTitle;
 	}
+
 	public void setNoticeTitle(String noticeTitle) {
 		this.noticeTitle = noticeTitle;
 	}
+
 	public String getNoticeContent() {
 		return noticeContent;
 	}
+
 	public void setNoticeContent(String noticeContent) {
 		this.noticeContent = noticeContent;
 	}
+
 	@Override
 	public String toString() {
 		return "NoticeUpdateCommand [noticeTitle=" + noticeTitle + ", noticeContent=" + noticeContent + ", noticeSeq="
 				+ noticeSeq + "]";
 	}
-	
-	
 
 }
