+++ qCali/src/main/java/com/group/exam/notice/service/NoticeService.java
@@ -8,10 +8,10 @@ import com.group.exam.utils.Criteria;
 
 public interface NoticeService {
 	public List<NoticeVo> list(Criteria cri);
-	public NoticeVo detail(int noticeSeq);
+	public NoticeVo detail(Long noticeSeq);
 	public void write(WriteCommand writeCommand);
-	public void delete(int noticeSeq);
-	public void update(String noticeTitle, String noticeContent, int noticeSeq);
+	public void delete(Long noticeSeq);
+	public void update(String noticeTitle, String noticeContent, Long noticeSeq);
 	public int noticeTotal();
 	
 }
